+++ mapper/lib/src/model/type_info.dart
@@ -1,3 +1,7 @@
+import 'package:reflectable/reflectable.dart';
+
+import './value_decorators.dart';
+
 /// Provides enhanced type information based on `Type.toString()` value
 class TypeInfo {
   Type type;
@@ -8,28 +12,36 @@ class TypeInfo {
 
   Type genericType;
 
+  Iterable<Type> parameters = []; // Type<T, K, V, etc.>
+
   bool isDynamic;
   bool isMap;
   bool isList;
   bool isSet;
   bool isIterable;
+  bool isEnum;
 
   TypeInfo(this.type);
 
   @override
   String toString() =>
-      'typeName: $typeName, scalarTypeName: $scalarTypeName, genericType: ' +
+      'typeName: $typeName, parameters: $parameters, genericType: ' +
       genericType.runtimeType.toString();
 }
 
 /// Abstract class for custom typeInfo decorator implementations
 abstract class ITypeInfoDecorator {
+  void init(Map<String, ClassMirror> knownClasses,
+      Map<Type, ValueDecoratorFunction> valueDecorators);
   TypeInfo decorate(TypeInfo typeInfo);
 }
 
 final defaultTypeInfoDecorator = DefaultTypeInfoDecorator();
 
 class DefaultTypeInfoDecorator implements ITypeInfoDecorator {
+  Map<String, ClassMirror> _knownClasses;
+  Iterable<Type> _valueDecoratorTypes;
+
   @override
   TypeInfo decorate(TypeInfo typeInfo) {
     final type = typeInfo.type;
@@ -45,8 +57,26 @@ class DefaultTypeInfoDecorator implements ITypeInfoDecorator {
     typeInfo.scalarType = detectScalarType(typeInfo);
     typeInfo.genericType = detectGenericType(typeInfo);
 
+    if (typeName != null && _knownClasses[typeName] != null) {
+      typeInfo.isEnum = _knownClasses[typeName].isEnum;
+    }
+
+    typeInfo.parameters =
+        getTypeParams(typeInfo).map((typeName) => detectTypeByName(typeName));
+
     return typeInfo;
   }
+
+  Iterable<String> getTypeParams(TypeInfo typeInfo) =>
+      typeInfo.typeName.indexOf('<') > 0
+          ? RegExp('<(.+)>')
+              .allMatches(typeInfo.typeName)
+              .first
+              .group(1)
+              .split(',')
+              .map((x) => x.trim())
+              .toList()
+          : [];
 
   String detectScalarTypeName(TypeInfo typeInfo) => typeInfo.isIterable
       ? RegExp('<(.+)>').allMatches(typeInfo.typeName).first.group(1)
@@ -65,10 +95,8 @@ class DefaultTypeInfoDecorator implements ITypeInfoDecorator {
     return null;
   }
 
-  Type detectScalarType(TypeInfo typeInfo) {
-    typeInfo.scalarTypeName = detectScalarTypeName(typeInfo);
-    if (typeInfo.isDynamic) return dynamic;
-    switch (typeInfo.scalarTypeName) {
+  Type detectTypeByName(String name) {
+    switch (name) {
       case 'DateTime':
         return DateTime;
       case 'num':
@@ -86,7 +114,24 @@ class DefaultTypeInfoDecorator implements ITypeInfoDecorator {
       case 'Symbol':
         return Symbol;
       default:
-        return null;
+        if (_knownClasses[name] != null) {
+          return _knownClasses[name].reflectedType;
+        }
+        return _valueDecoratorTypes.firstWhere((t) => t.toString() == name,
+            orElse: () => dynamic);
     }
+  }
+
+  Type detectScalarType(TypeInfo typeInfo) {
+    typeInfo.scalarTypeName = detectScalarTypeName(typeInfo);
+    if (typeInfo.isDynamic) return dynamic;
+    return detectTypeByName(typeInfo.scalarTypeName);
+  }
+
+  @override
+  void init(Map<String, ClassMirror> knownClasses,
+      Map<Type, ValueDecoratorFunction> valueDecorators) {
+    _knownClasses = knownClasses;
+    _valueDecoratorTypes = valueDecorators.keys;
   }
 }